# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hakimi Healthcare API
  version: 1.0.0
  license:
    name: UNLICENSED
    identifier: UNLICENSED
  description: |
    REST API for Hakimi / ሀኪሜ, a healthcare appointment platform for Ethiopia.

    Notes:
    - Workforce API operations require an OIDC bearer access token with workforce MFA.
    - Authorization is default deny and derives roles and facility scope from server-side PostgreSQL state.
    - Patient authentication and self-service remain unavailable.
    - `DELETE` on facilities, practitioners, and assignments performs a soft delete.
    - Practitioner facility primary reassignment is handled atomically inside one transaction.
servers:
  - url: http://127.0.0.1:3001
security:
  - workforceBearer: []
...
paths:
...
  /api/v1/appointments:
    post:
      tags: [Appointments]
      summary: Create an appointment
      description: |
        Creates an appointment in a single transaction after confirming the facility, practitioner, and patient
        are active and linked correctly. Only non-cancelled appointments block a practitioner time slot.
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentInput'
            examples:
              schedule:
                summary: Schedule a new appointment
                value:
                  patientId: 11111111-1111-4111-8111-111111111111
                  practitionerId: 22222222-2222-4222-8222-222222222222
                  facilityId: 33333333-3333-4333-8333-333333333333
                  scheduledStart: '2026-08-07T09:00:00+03:00'
                  scheduledEnd: '2026-08-07T09:30:00+03:00'
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '201':
          description: Created appointment
          headers:
            Location:
              description: URL of the created appointment
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/AppointmentCreateLookupError'
        '409':
          $ref: '#/components/responses/AppointmentCreateConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Appointments]
      summary: List appointments
      description: |
        Returns appointments in deterministic scheduled-start order. Date range filters use overlapping windows so
        appointments partially inside the requested window are included.
      operationId: listAppointments
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: facilityId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: practitionerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: patientId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
        - name: from
          in: query
          required: false
          description: Inclusive lower bound for overlapping appointment windows.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Exclusive upper bound for overlapping appointment windows.
          schema:
            type: string
            format: date-time
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Paged appointment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/appointments/{appointmentId}:
    parameters:
      - name: appointmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Appointments]
      summary: Get an appointment by id
      operationId: getAppointmentById
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/AppointmentNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Appointments]
      summary: Update an appointment
      description: |
        Allows rescheduling and state transitions from SCHEDULED to CONFIRMED or from CONFIRMED to COMPLETED or NO_SHOW.
        CANCELLED appointments remain terminal and use the cancel endpoint instead.
      operationId: updateAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentInput'
            examples:
              confirm:
                summary: Confirm an appointment
                value:
                  status: CONFIRMED
              reschedule:
                summary: Reschedule an appointment
                value:
                  scheduledStart: '2026-08-07T10:00:00+03:00'
                  scheduledEnd: '2026-08-07T10:30:00+03:00'
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/AppointmentNotFound'
        '409':
          $ref: '#/components/responses/AppointmentPatchConflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/appointments/{appointmentId}/cancel:
    parameters:
      - name: appointmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Appointments]
      summary: Cancel an appointment
      description: |
        Marks an appointment as cancelled without deleting it. Repeating the request for an already cancelled
        appointment is idempotent and returns the existing record.
      operationId: cancelAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelAppointmentInput'
            examples:
              cancel:
                summary: Cancel an appointment
                value:
                  cancellationReason: Patient requested a later time
      responses:
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

        '200':
          description: Cancelled appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/AppointmentNotFound'
        '409':
          $ref: '#/components/responses/AppointmentStateConflict'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    workforceBearer:
      type: oauth2
      description: |
        Workforce-only OIDC Authorization Code with PKCE. The resource server
        requires a short-lived signed JWT access token, approved workforce MFA,
        and an active server-provisioned actor, role, scope, and local session.
        Patient accounts and role-administration APIs are not available.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.example.com/oauth2/authorize
          tokenUrl: https://identity.example.com/oauth2/token
          scopes: {}
  schemas:
    FacilityType:
      type: string
      enum:
        - hospital
        - clinic
        - health_center
        - diagnostic_center
        - pharmacy
...
    AdministrativeSex:
      type: string
      enum:
        - female
        - male
        - other
        - unknown
...
    AppointmentStatus:
      type: string
      enum:
        - SCHEDULED
        - CONFIRMED
        - COMPLETED
        - CANCELLED
        - NO_SHOW
    AppointmentFacilitySummary:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - name
        - facilityType
        - region
        - city
        - isActive
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        facilityType:
          $ref: '#/components/schemas/FacilityType'
        region:
          type: string
        city:
          type: string
        isActive:
          type: boolean
    AppointmentPatientSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - firstName
        - middleName
        - lastName
        - dateOfBirth
        - administrativeSex
        - isActive
        - medicalRecordNumber
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        middleName:
          type:
            - string
            - 'null'
        lastName:
          type:
            - string
            - 'null'
        dateOfBirth:
          type:
            - string
            - 'null'
          format: date
        administrativeSex:
          $ref: '#/components/schemas/AdministrativeSex'
        isActive:
          type: boolean
        medicalRecordNumber:
          type: string
    AppointmentPractitionerSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - code
        - firstName
        - middleName
        - lastName
        - profession
        - isActive
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        firstName:
          type: string
        middleName:
          type:
            - string
            - 'null'
        lastName:
          type: string
        profession:
          type: string
        isActive:
          type: boolean
    Appointment:
      type: object
      additionalProperties: false
      required:
        - id
        - patientId
        - practitionerId
        - facilityId
        - scheduledStart
        - scheduledEnd
        - status
        - cancellationReason
        - cancelledAt
        - createdAt
        - updatedAt
        - patient
        - practitioner
        - facility
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        practitionerId:
          type: string
          format: uuid
        facilityId:
          type: string
          format: uuid
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        cancellationReason:
          type:
            - string
            - 'null'
        cancelledAt:
          type:
            - string
            - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        patient:
          $ref: '#/components/schemas/AppointmentPatientSummary'
        practitioner:
          $ref: '#/components/schemas/AppointmentPractitionerSummary'
        facility:
          $ref: '#/components/schemas/AppointmentFacilitySummary'
    CreateAppointmentInput:
      type: object
      additionalProperties: false
      required:
        - patientId
        - practitionerId
        - facilityId
        - scheduledStart
        - scheduledEnd
      properties:
        patientId:
          type: string
          format: uuid
        practitionerId:
          type: string
          format: uuid
        facilityId:
          type: string
          format: uuid
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
    UpdateAppointmentInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
    CancelAppointmentInput:
      type: object
      additionalProperties: false
      required:
        - cancellationReason
      properties:
        cancellationReason:
          type: string
          minLength: 1
          maxLength: 1000
    AppointmentPagination:
      type: object
      additionalProperties: false
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    AppointmentListResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        pagination:
          $ref: '#/components/schemas/AppointmentPagination'
    ApiErrorDetail:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - INVALID_JSON
                - AUTHENTICATION_REQUIRED
                - FORBIDDEN
                - FACILITY_NOT_FOUND
                - FACILITY_CODE_CONFLICT
                - FACILITY_LICENSE_CONFLICT
                - FACILITY_INACTIVE
                - PRACTITIONER_NOT_FOUND
                - PRACTITIONER_CODE_CONFLICT
                - PRACTITIONER_LICENSE_CONFLICT
                - INACTIVE_PRACTITIONER
                - INACTIVE_FACILITY
                - ASSIGNMENT_NOT_FOUND
                - ASSIGNMENT_CONFLICT
                - PATIENT_NOT_FOUND
                - INACTIVE_PATIENT
                - PATIENT_REGISTRATION_NOT_FOUND
                - PATIENT_REGISTRATION_CONFLICT
                - APPOINTMENT_NOT_FOUND
                - APPOINTMENT_CONFLICT
                - APPOINTMENT_STATE_CONFLICT
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: array
              items:
                $ref: '#/components/schemas/ApiErrorDetail'
  responses:
    AuthenticationRequired:
      description: Missing, invalid, expired, revoked, or otherwise ineligible workforce authentication
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            authenticationRequired:
              value:
                error:
                  code: AUTHENTICATION_REQUIRED
                  message: Authentication required
    Forbidden:
      description: Authenticated workforce actor is not permitted to perform this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            forbidden:
              value:
                error:
                  code: FORBIDDEN
                  message: Forbidden
    ValidationError:
      description: Validation failure or malformed request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            validation:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Request validation failed
                  details:
                    - field: name
                      message: Name is required
...
    AppointmentNotFound:
      description: Appointment not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            notFound:
              value:
                error:
                  code: APPOINTMENT_NOT_FOUND
                  message: Appointment not found
...
    AppointmentStateConflict:
      description: Appointment state does not allow the requested operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            invalidState:
              value:
                error:
                  code: APPOINTMENT_STATE_CONFLICT
                  message: Appointment cannot be modified in its current state
    AppointmentCreateLookupError:
      description: Appointment creation can fail because one of the linked records is missing.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: FACILITY_NOT_FOUND
                      message:
                        type: string
                        const: Facility not found
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: PRACTITIONER_NOT_FOUND
                      message:
                        type: string
                        const: Practitioner not found
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: PATIENT_NOT_FOUND
                      message:
                        type: string
                        const: Patient not found
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: PATIENT_REGISTRATION_NOT_FOUND
                      message:
                        type: string
                        const: Patient is not registered at this facility
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: ASSIGNMENT_NOT_FOUND
                      message:
                        type: string
                        const: Assignment not found
          examples:
            facilityMissing:
              summary: Facility not found
              value:
                error:
                  code: FACILITY_NOT_FOUND
                  message: Facility not found
            practitionerMissing:
              summary: Practitioner not found
              value:
                error:
                  code: PRACTITIONER_NOT_FOUND
                  message: Practitioner not found
            patientMissing:
              summary: Patient not found
              value:
                error:
                  code: PATIENT_NOT_FOUND
                  message: Patient not found
            registrationMissing:
              summary: Patient not registered at the facility
              value:
                error:
                  code: PATIENT_REGISTRATION_NOT_FOUND
                  message: Patient is not registered at this facility
            assignmentMissing:
              summary: Practitioner is not assigned to the facility
              value:
                error:
                  code: ASSIGNMENT_NOT_FOUND
                  message: Assignment not found
    AppointmentCreateConflict:
      description: Appointment creation can fail because the facility, practitioner, or patient is inactive, or because the practitioner is already booked.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: FACILITY_INACTIVE
                      message:
                        type: string
                        const: Facility is inactive
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: INACTIVE_PRACTITIONER
                      message:
                        type: string
                        const: Practitioner is inactive
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: INACTIVE_PATIENT
                      message:
                        type: string
                        const: Patient is inactive
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: APPOINTMENT_CONFLICT
                      message:
                        type: string
                        const: Appointment conflicts with an existing appointment
          examples:
            facilityInactive:
              summary: Facility is inactive
              value:
                error:
                  code: FACILITY_INACTIVE
                  message: Facility is inactive
            practitionerInactive:
              summary: Practitioner is inactive
              value:
                error:
                  code: INACTIVE_PRACTITIONER
                  message: Practitioner is inactive
            patientInactive:
              summary: Patient is inactive
              value:
                error:
                  code: INACTIVE_PATIENT
                  message: Patient is inactive
            appointmentConflict:
              summary: Practitioner already booked
              value:
                error:
                  code: APPOINTMENT_CONFLICT
                  message: Appointment conflicts with an existing appointment
    AppointmentPatchConflict:
      description: Appointment patch can fail because the requested state transition is invalid or the new time overlaps another appointment.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: APPOINTMENT_CONFLICT
                      message:
                        type: string
                        const: Appointment conflicts with an existing appointment
              - type: object
                additionalProperties: false
                required: [error]
                properties:
                  error:
                    type: object
                    additionalProperties: false
                    required: [code, message]
                    properties:
                      code:
                        type: string
                        const: APPOINTMENT_STATE_CONFLICT
                      message:
                        type: string
                        const: Appointment cannot be modified in its current state
          examples:
            conflict:
              summary: Appointment overlaps another appointment
              value:
                error:
                  code: APPOINTMENT_CONFLICT
                  message: Appointment conflicts with an existing appointment
            stateConflict:
              summary: Invalid appointment transition
              value:
                error:
                  code: APPOINTMENT_STATE_CONFLICT
                  message: Appointment cannot be modified in its current state
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            internal:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: Internal server error